+++ src/Entity/Net/Variables.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Extraton\TonClient\Entity\Net;
+
+use Extraton\TonClient\Entity\Params;
+use Extraton\TonClient\Exception\DataException;
+
+/**
+ * Query Variables
+ */
+class Variables implements Params
+{
+
+    public function jsonSerialize(): array
+    {
+        // TODO: Implement jsonSerialize() method.
+    }
+}