+++ tests/Unit/NetTest.php
@@ -4,9 +4,13 @@ declare(strict_types=1);
 
 namespace Extraton\Tests\Unit\TonClient;
 
+use Extraton\TonClient\Entity\Crypto\ResultOfModularPower;
+use Extraton\TonClient\Entity\Net\EndpointsSet;
 use Extraton\TonClient\Entity\Net\Filters;
 use Extraton\TonClient\Entity\Net\OrderBy;
 use Extraton\TonClient\Entity\Net\QueryInterface;
+use Extraton\TonClient\Entity\Net\ResultOfFindLastShardBlock;
+use Extraton\TonClient\Entity\Net\ResultOfQuery;
 use Extraton\TonClient\Entity\Net\ResultOfQueryCollection;
 use Extraton\TonClient\Entity\Net\ResultOfSubscribeCollection;
 use Extraton\TonClient\Entity\Net\ResultOfWaitForCollection;
@@ -268,5 +272,136 @@ class NetTest extends AbstractModuleTest
             ->willReturn($this->mockPromise);
 
         $this->net->unsubscribe($handle);
+    }
+
+    /**
+     * @covers ::query
+     */
+    public function testQuerySuccessResult(): void
+    {
+        $query = uniqid(microtime(), true);
+        $variables = [
+            uniqid(microtime(), true) => uniqid(microtime(), true)
+        ];
+
+        $response = new Response(
+            [
+                uniqid(microtime(), true)
+            ]
+        );
+
+        $this->mockPromise->expects(self::once())
+            ->method('wait')
+            ->with()
+            ->willReturn($response);
+
+        $this->mockTonClient->expects(self::once())
+            ->method('request')
+            ->with(
+                'net.query',
+                [
+                    'query'     => $query,
+                    'variables' => $variables,
+                ]
+            )
+            ->willReturn($this->mockPromise);
+
+        $expected = new ResultOfQuery($response);
+
+        self::assertEquals($expected, $this->net->query($query, $variables));
+    }
+
+    /**
+     * @covers ::findLastShardBlock
+     */
+    public function testFindLastShardBlockSuccessResult(): void
+    {
+        $address = uniqid(microtime(), true);
+
+        $response = new Response(
+            [
+                uniqid(microtime(), true)
+            ]
+        );
+
+        $this->mockPromise->expects(self::once())
+            ->method('wait')
+            ->with()
+            ->willReturn($response);
+
+        $this->mockTonClient->expects(self::once())
+            ->method('request')
+            ->with(
+                'net.find_last_shard_block',
+                [
+                    'address' => $address,
+                ]
+            )
+            ->willReturn($this->mockPromise);
+
+        $expected = new ResultOfFindLastShardBlock($response);
+
+        self::assertEquals($expected, $this->net->findLastShardBlock($address));
+    }
+
+    /**
+     * @covers ::setEndpoints
+     */
+    public function testSetEndpointsSuccessResult(): void
+    {
+        $endpoints = [
+            uniqid(microtime(), true) => uniqid(microtime(), true),
+        ];
+
+        $response = new Response(
+            [
+                uniqid(microtime(), true)
+            ]
+        );
+
+        $this->mockPromise->expects(self::once())
+            ->method('wait')
+            ->with()
+            ->willReturn($response);
+
+        $this->mockTonClient->expects(self::once())
+            ->method('request')
+            ->with(
+                'net.set_endpoints',
+                [
+                    'endpoints' => $endpoints,
+                ]
+            )
+            ->willReturn($this->mockPromise);
+
+        $this->net->setEndpoints($endpoints);
+    }
+
+    /**
+     * @covers ::fetchEndpoints
+     */
+    public function testFetchEndpointsSuccessResult(): void
+    {
+        $response = new Response(
+            [
+                uniqid(microtime(), true)
+            ]
+        );
+
+        $this->mockPromise->expects(self::once())
+            ->method('wait')
+            ->with()
+            ->willReturn($response);
+
+        $this->mockTonClient->expects(self::once())
+            ->method('request')
+            ->with(
+                'net.fetch_endpoints'
+            )
+            ->willReturn($this->mockPromise);
+
+        $expected = new EndpointsSet($response);
+
+        self::assertEquals($expected, $this->net->fetchEndpoints());
     }
 }